Delete README.md
After performing EDA in the data, I carried out random forest in the data.
This is the result of running a Random Forest Classifier model on the churn dataset. The model is defined with the default parameters and it was fit on the data with the training set and used to make predictions on the test set.

The feature importances are listed, which show the relative importance of each feature in predicting churn. The values are scaled between 0 and 1, with higher values indicating greater importance.

Based on the results, the most important features for predicting churn in this dataset are Age, NumOfProducts, Balance, and EstimatedSalary. The least important features are HasCrCard, Spain, and IsActiveMember.